example/requirements-gcp.txt
google-auth
google-cloud-storage
cloud_auth_tpm[gcp]
tpm2-pytss
cryptography